% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_covreg.R
\name{cv.mr_covreg}
\alias{cv.mr_covreg}
\title{Cross-validation for mr_covreg}
\usage{
cv.mr_covreg(bx, bw, by, S, lambda = numeric(0), nlam = 100,
  nfolds = 10)
}
\arguments{
\item{bx}{Vector of estimates of the genetic variant-risk factor associations.}

\item{bw}{Matrix of estimates of the genetic variant-covariate associations
estimates. The jth column of the matrix is a vector of the estimates of the
genetic variant associations with the jth covariate.}

\item{by}{Vector of estimates of the genetic variant-outcome associations.}

\item{S}{Diagonal matrix where the jth diagonal entry is the inverse of the
variance of the jth genetic variant-outcome association estimate.}

\item{lambda}{Sequence of lambda values to be used in cross-validation. If not
specified (which is the default setting), the sequence of lambda values is
chosen by the glmnet package.}

\item{nlam}{Number of lambda values to use in cross-validation. Default is 100.
Note, if a lambda sequence is given, this parameter is redundant.}

\item{nfolds}{Number of folds for cross-validation. Default is 10.}
}
\value{
\item{glmnet.fit}{List containing a matrix of coefficients and a vector of
   the number of non-zero coefficients.}
 \item{lamseq}{Sequence of lambda values used in cross-validation.}
 \item{lambda.min}{The value of lambda that minimised the test mean squared
   error.}
 \item{lambda.min}{The value of lambda that minimised the test mean squared
   error with the 1 standard deviation rule applied.}
}
\description{
Implements K-fold cross-validation for mr_covreg where the target function is
the mean squared error.
}
