% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_covreg.R
\name{mr_covreg}
\alias{mr_covreg}
\title{Causal effect estimation via covariate regularization}
\usage{
mr_covreg(bx, bw, by, S, klessp = TRUE, lambda = numeric(0),
  nlam = 100, K = 10, cv_mt = 2)
}
\arguments{
\item{bx}{Vector of estimates of the genetic variant-risk factor associations.}

\item{bw}{Matrix of estimates of the genetic variant-covariate associations
estimates. The jth column of the matrix is a vector of the estimates of the
genetic variant associations with the jth covariate.}

\item{by}{Vector of estimates of the genetic variant-outcome associations.}

\item{S}{Diagonal matrix where the jth diagonal entry is the inverse of the
variance of the jth genetic variant-outcome association estimate.}

\item{klessp}{Indicates whether the tuning parameter should be always
sufficiently large such that there are always less than p - 1 covariates
with a non-zero coefficient.}

\item{lambda}{Sequence of lambda values to be used in cross-validation. If not
specified (which is the default setting), the sequence of lambda values is
chosen by the glmnet package.}

\item{nlam}{Number of lambda values to use in cross-validation. Default is 100.
Note, if a lambda sequence is given, this parameter is redundant.}

\item{K}{Number of folds for cross-validation. Default is 10.}

\item{cv_mt}{Controls which target function to use in cross-validation. If set
at 1, the tuning parameter is selected independent of the
genetic variant-risk factor associations. Otherwise, the target function is
the mean squared error (which is the default).}
}
\value{
\item{thest}{Causal effect estimate.}
 \item{thest_1se}{Causal effect estimate with the 1 standard error rule
   applied.}
 \item{a}{Regularized covariate cofficients.}
 \item{a_1se}{Regularized covariate cofficients with the 1 standard error
   rule applied.}
 \item{lambda}{Value of lambda chosen by cross-validation.}
 \item{lambda_1se}{Value of lambda chosen by cross-validation with the 1
   standard error rule applied.}
 \item{lamseq}{Sequence of lambda values used in cross-validation.}
}
\description{
Estimates a causal effect by implementing regularization on potential
pleiotropic covariates. The tuning parameter is chosen by cross-validation.
}
