% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_covreg.R
\name{mr_covreg_lam}
\alias{mr_covreg_lam}
\title{Causal effect estimation via covariate regularization for a specified tuning
parameter value}
\usage{
mr_covreg_lam(bx, bw, by, S, lambda)
}
\arguments{
\item{bx}{Vector of estimates of the genetic variant-risk factor associations.}

\item{bw}{Matrix of estimates of the genetic variant-covariate associations
estimates. The jth column of the matrix is a vector of the estimates of the
genetic variant associations with the jth covariate.}

\item{by}{Vector of estimates of the genetic variant-outcome associations.}

\item{S}{Diagonal matrix where the jth diagonal entry is the inverse of the
variance of the jth genetic variant-outcome association estimate.}
}
\value{
\item{thest}{Causal effect estimate.}
 \item{a}{Regularized covariate cofficients.}
}
\description{
Estimates a causal effect by implementing regularization on potential
pleiotropic covariates for a given value of the tuning parameter.
}
